requirements.txt
paho_mqtt>=2.1.0
cryptography>=44.0.0
